requirements.txt
Flask==3.1.0
ipython==8.31.0
requests==2.32.3
bcrypt==4.2.1
Flask-MySQL==1.6.0
PyMySQL==1.1.1
SQLAlchemy==2.0.37
python-dotenv==1.0.1
alembic==1.14.1
